+++ drizzle/0008_smart_stingray.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `song` ADD `mix_url` text;--> statement-breakpoint
+ALTER TABLE `song` ADD `mix_key` text;